backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Caching==2.1.0
python-multipart==0.0.6
Werkzeug==3.0.1

# Deep Learning & NLP
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Traditional ML
scikit-learn==1.3.2
xgboost==2.0.1
pandas==2.1.3
numpy==1.25.2

# Text Processing
PyPDF2==3.0.1
python-docx==0.8.11
textract==1.6.5

# Utilities
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
python-dotenv==1.0.0
joblib==1.3.2
